// src/model/gym.model.ts
export interface IGymModel {
    id: number
    region: string
    name: string
    address: string
    detailAddress: string
    size: string
    operatingOrganization: string
    phoneNumber: string
    operatingTimeDay: string
    operatingTimeWeekend: string
    operatingTImeHoliday: string
    rentalStatus: boolean
    price: string
    parkingInfo: string
    homepage: string
    sportsType: string
    currentStatus: string
    facilities: string
    etc: string
}